Build base URL instead of using window.location
Using window.location to build the base URL will give funny
results when query args are passed in (in the case of tracking
metrics), since they are included. This means the following lines

  var apiUrl = window.location + "api/";

would evaluate to:

  apiURL = https://predictive-market-stress-testing.mybluemix.net/?foo=barapi/

Which led to several API calls hanging.

Build the base URL correctly with protocol, host and path instead

  apiURL = location.protocol + '//' + location.host + location.pathname